+++ src/main/java/com/example/demo/service/impl/SaleRecordServiceImpl.java
@@ -30,7 +30,7 @@ public class SaleRecordServiceImpl implements SaleRecordService {
     @Override
     public List<SaleRecord> findSaleRecordByQuerys(String merchantName, String businessCode, String recodeUsername,
                                                    String uniqueCode, String productDaleiName, String productZhongleiName,
-                                                   Long startTime, Long endTime, Pageable pageable) {
+                                                   Integer startTime, Integer endTime, Pageable pageable) {
         return saleRecordRepository.findSaleRecordByQuerys(merchantName, businessCode, recodeUsername, uniqueCode,
                 productDaleiName, productZhongleiName, startTime, endTime, pageable);
     }
